Merging all the sql agent runner project into main solution and then taking the classes from sqlagentcore and putting them in the services project adding a log of tests and ignoring some rhino bus related ones as weird stuff happens when they all run together...it cant find the rhino.esb section sometimes maybe got to do with webconfigs in different projects? Im not sure.
Also added a special locking section to the log4net.xml appender section
to ensure the file is only locked for the duration of a message being written to
the log. This is to prevent file locking going on for too long and then causing write contentions
with other custom logger instances that may be trying to write to the same file
at the same time.