Added automapper to map TaskRunnerPropertyModel to PropertyWithValue in the taskrunner.core for sending messages. this allows us to handle only models with our interaction between client and server but when we need to map it to another type its painless and only at the point of execution.
Added my old MapFrom and MapCutom classes for making automapper automatic.
However in this case since we are mapping a type in our dll to a type in another dll
we have to do it the old fashioned way and call the static register map method in the global.asax
This is because MapFrom and MapCutom only map to a type from another type
but in this instance we want to map from our type to another type.

I will look into writing another method and interface to handle this scenario.
Fixed some tests.. dont know where I was going with all that extra serialization last night.
However I did add two nice new overloaded methods for wrapping json.net serialization with types and string
rather than using files.