Removing redundant sql sprocs....as they are only generated during tests anyway. Refactored some of sql script runner core code but its a bit of a mess... Will need a lot of work.
Added some directorymanager class for simply wrapping processes that required the creation of
a directory and then dispose/deletes it when finished. Also checks to see if the directory already
exists and if so throws an error . You can tell it to force deletion... Its a bit of an abuse of Idisposable
but hey lazy programming ftw.