Adding structuremap ioc container to automatically handle dependency injection. Using registries and dependency resolvers using configured containers etc. Moved things to Core as it makes it easier to test and more object oriented approach.
Trying to manage the layers correctly having pocos and models in business logic
and managers and helpers in core. Everything else should just have what they need for
their application to run. Not exactly fond of have TaskRunnerReflector in core but
in order to inject it using a registry in common it couldnt be in Taskrunner.core
as this would have lead to a circular dependency. Perhaps registries shouldnt be in Common
but I like them there..I think it might work we shall see.

New branch as this is a big merge and many things might have broken unbeknownst to me.